# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/login": {
            "post": {
                "description": "login",
                "consumes": [
                    "application/x-www-form-urlencoded",
                    "multipart/form-data",
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "login",
                "operationId": "api-1-login-post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Password",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apis.RespDone"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apis.RespFail"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apis.RespFail"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apis.RespFail"
                        }
                    }
                }
            }
        },
        "/api/password": {
            "post": {
                "description": "change password",
                "consumes": [
                    "application/x-www-form-urlencoded",
                    "multipart/form-data",
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Change password",
                "operationId": "api-1-password-post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Old Password",
                        "name": "old_password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "New Password",
                        "name": "new_password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Confirm Password",
                        "name": "password_confirm",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apis.RespDone"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apis.RespFail"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apis.RespFail"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apis.RespFail"
                        }
                    }
                }
            }
        },
        "/api/password/forgot": {
            "post": {
                "description": "forgot password",
                "consumes": [
                    "application/x-www-form-urlencoded",
                    "multipart/form-data",
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Forgot password",
                "operationId": "api-1-password-forgot-post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Login name",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Mobile number",
                        "name": "mobile",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Email address",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apis.RespDone"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apis.RespFail"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apis.RespFail"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apis.RespFail"
                        }
                    }
                }
            }
        },
        "/api/password/reset": {
            "post": {
                "description": "reset password, form:rt, json:token",
                "consumes": [
                    "application/x-www-form-urlencoded",
                    "multipart/form-data",
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Reset password",
                "operationId": "api-1-password-reset-post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Login name",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Password",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Confirm Password",
                        "name": "password_confirm",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "rt",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apis.RespDone"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apis.RespFail"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apis.RespFail"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apis.RespFail"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "apis.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "apis.RespDone": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "main data"
                },
                "extra": {
                    "description": "extra data"
                },
                "ok": {
                    "description": "OK",
                    "type": "boolean"
                }
            }
        },
        "apis.RespFail": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/apis.Error"
                },
                "ok": {
                    "type": "boolean"
                }
            }
        }
    }
}